Added multiport mode, enjoy (2)
