--- Models/ArInternalMetadatum.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace ArmstrongServer.Models;
-
-public partial class ArInternalMetadatum
-{
-    public string Key { get; set; } = null!;
-
-    public string? Value { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-
-    public DateTime UpdatedAt { get; set; }
-}
