+++ Models/ConfigModels/ServerProps.cs
@@ -1,0 +1,12 @@
+namespace ArmstrongServer.Models.ConfigModels
+{
+  public class ServerProps
+  {
+    public int ServerId { get; set; }
+    public int DeadPollingTime { get; set; }
+    public int PollingTimeout { get; set; }
+    public int MinimalPollingTimeout { get; set; }
+    public string ComPortName { get; set; }
+    public int ComPortBaudRate { get; set; }
+  }
+}
