--- Models/DeviceRegGroup.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace ArmstrongServer.Models;
-
-public partial class DeviceRegGroup
-{
-    public long Id { get; set; }
-
-    public string? Name { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-
-    public DateTime UpdatedAt { get; set; }
-
-    public virtual ICollection<Device> Devices { get; set; } = new List<Device>();
-}
