--- Models/MeasurementGroup.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace ArmstrongServer.Models;
-
-public partial class MeasurementGroup
-{
-    public long Id { get; set; }
-
-    public string? Name { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-
-    public DateTime UpdatedAt { get; set; }
-
-    public virtual ICollection<DeviceModel> DeviceModels { get; set; } = new List<DeviceModel>();
-
-    public virtual ICollection<MeasurementClass> MeasurementClasses { get; set; } = new List<MeasurementClass>();
-}
