--- Models/MultiportCollection.cs
@@ -1,8 +1,0 @@
-namespace ArmstrongServer.Models
-{
-  public class MultiportCollection
-  {
-    public ServerAttributes ServerAttributes { get; set; }
-    public PortSettings PortSettings { get; set; }
-  }
-}
