+++ Models/Server.cs
@@ -1,29 +1,97 @@
-﻿using System;
-using System.Collections.Generic;
+using ArmstrongServer.Models.DataModels;
+using ArmstrongServer.Models.ConfigModels;
+using ArmstrongServer.Data;
+using Microsoft.EntityFrameworkCore;
+using ArmstrongServer.Interfaces;
 
-namespace ArmstrongServer.Models;
-
-public partial class Server
+namespace ArmstrongServer.Models
 {
-    public long Id { get; set; }
+  public class Server : IServer<ArmsWebappDevelopmentContext>
+  {
+    public int ServerId { get; set; }
+    public int DeadPollingTime { get; set; }
+    public int PollingTimeout { get; set; }
+    public int MinimalPollingTimeout { get; set; }
+    public ComPort ServerComPort { get; set; }
+    public IOrderedQueryable<Channel> Channels { get; set; }
+    public IOrderedQueryable<History> Histories { get; set; }
+    public ArmsWebappDevelopmentContext DataContext { get; set; }
 
-    public string? Name { get; set; }
+    private bool _isPolling;
 
-    public string? IpAdress { get; set; }
+    public Server(ServerProps serverProps)
+    {
+      ServerId = serverProps.ServerId;
+      ServerComPort = new(serverProps);
+      DeadPollingTime = serverProps.DeadPollingTime;
+      PollingTimeout = serverProps.PollingTimeout;
+      MinimalPollingTimeout = serverProps.MinimalPollingTimeout;
 
-    public int? InventoryId { get; set; }
+      DataContext = new();
 
-    public long ServiceId { get; set; }
+      Channels = DataContext
+        .Channels.Where(c => c.ServerId == this.ServerId)
+        .Include(x => x.Device.DeviceModel).Include(x => x.Room).Include(x => x.Device.DeviceModel.MeasurementClass)
+        .OrderBy(x => x.ChannelId);
 
-    public long RoomId { get; set; }
+      foreach (var channel in Channels)
+        channel.Initialization(serverProps);
+    }
 
-    public DateTime CreatedAt { get; set; }
+    public void Start()
+    {
+      _isPolling = true;
+      Console.WriteLine($"Thread for Server {ServerId} started.");
 
-    public DateTime UpdatedAt { get; set; }
+      var channelsCount = Channels.Count();
+      var summPollingPortOverhead = channelsCount * this.DeadPollingTime * 2;
+      var realPollingTimeout = this.PollingTimeout > summPollingPortOverhead
+        ? (int)(this.PollingTimeout - summPollingPortOverhead)
+        : this.MinimalPollingTimeout;
 
-    public virtual ICollection<Channel> Channels { get; set; } = new List<Channel>();
+      while (_isPolling)
+      {
+        foreach (var c in Channels)
+        {
+          c.StartOneshotDialogSession();
+        }
 
-    public virtual Room Room { get; set; } = null!;
+        foreach (var c in Channels)
+        {
+          if (c.EventErrorCount == 0)
+          {
+            var history = new History
+            {
+              ChannelId = c.Id,
+              EventSystemValue = c.EventSystemValue,
+              EventNotSystemValue = c.EventNotSystemValue,
+              EventImpulseValue = c.EventImpulseValue,
+              EventDatetime = c.EventDatetime,
+              UpdatedAt = c.UpdatedAt,
+              CreatedAt = DateTime.UtcNow
+            };
+            DataContext.Histories.Add(history);
+          }
+        }
 
-    public virtual Service Service { get; set; } = null!;
+        DataContext.SaveChanges();
+        Thread.Sleep(realPollingTimeout);
+
+        Console.WriteLine($"Server {ServerId} -- Work");
+        Task.Delay(1000).Wait(); // Пауза между опросами
+      }
+
+      Console.WriteLine($"Server {ServerId} -- Reload");
+    }
+
+    public void Stop()
+    {
+      _isPolling = false;
+    }
+
+    public void Update()
+    {
+      throw new NotImplementedException();
+    }
+  }
 }
