+++ Program.cs
@@ -1,59 +1,48 @@
-﻿using ArmstrongServer.Models;
-using ArmstrongServer.Data;
+﻿using ArmstrongServer.Models.ConfigModels;
+using ArmstrongServer.Models;
 using ArmstrongServer.Helpers;
-using Microsoft.EntityFrameworkCore;
 
 namespace ArmstrongServer
 {
   public static class Program
   {
-    public static void Main(string[] args)
+    public static List<Server> Servers { get; set; }
+    public static async Task Main(string[] args)
+    {
+      Initialization();
+
+      var serverThreads = new List<Thread>();
+
+      foreach (var server in Servers)
+      {
+        var thread = new Thread(() => server.Start());
+        serverThreads.Add(thread);
+        thread.Start();
+      }
+
+      Console.WriteLine("Starting server threads...");
+
+      Console.WriteLine("Press Enter to exit.");
+      Console.ReadLine();
+
+      Console.WriteLine("Stopping server threads...");
+
+      // this is not supported by the netcore 6, use concellationToken
+      foreach (var thread in serverThreads)
+      {
+        thread.Abort();
+      }
+
+      await Task.WhenAll(serverThreads.Select(thread => Task.Run(() => thread.Join())).ToArray());
+
+      Console.WriteLine("All server threads have completed.");
+    }
+
+    public static void Initialization()
     {
       AppSettings.Initialization();
-
+      Servers = ServerHelper.GetServerList(AppSettings.AppGeneralSettings.ServerProps);
       SayHelloWorld.Say();
-
-      var context = new ArmsWebappDevelopmentContext();
-
-      var channels = context
-        .Channels.Where(c => c.ServerId == AppSettings.AppServerAttributes.Id)
-        .Include(x => x.Device.DeviceModel).Include(x => x.Room).Include(x => x.Device.DeviceModel.MeasurementClass)
-        .OrderBy(x => x.ChannelId);
-
-      var channelsCount = channels.Count();
-      var summPollingPortOverhead = channelsCount * AppSettings.AppPortSettings.DeadPollingTime * 2;
-      var realPollingTimeout = AppSettings.AppGeneralSettings.ChannelPolingTimeout > summPollingPortOverhead
-        ? (int)(AppSettings.AppGeneralSettings.ChannelPolingTimeout - summPollingPortOverhead)
-        : AppSettings.AppPortSettings.MinimalPollingTimeout;
-
-      while (true)
-      {
-        foreach (var c in channels)
-        {
-          c.StartOneshotDialogSession();
-        }
-
-        foreach (var c in channels)
-        {
-          if (c.EventErrorCount == 0)
-          {
-            var history = new History
-            {
-              ChannelId = c.Id,
-              EventSystemValue = c.EventSystemValue,
-              EventNotSystemValue = c.EventNotSystemValue,
-              EventImpulseValue = c.EventImpulseValue,
-              EventDatetime = c.EventDatetime,
-              UpdatedAt = c.UpdatedAt,
-              CreatedAt = DateTime.UtcNow
-            };
-            context.Histories.Add(history);
-          }
-        }
-
-        context.SaveChanges();
-        Thread.Sleep(realPollingTimeout);
-      }
     }
   }
 }
