+++ appsettings.json
@@ -1,53 +1,30 @@
 {
   "GeneralSettings": {
     "AuthToken": "token",
-    "ChannelPolingTimeout": 40000,
     "ChannelPollingErrorCountLimit": 60,
-    "MultiportMode": false,
-    "MultiportCollection": [
+    "ServerProps": [
       {
-        "PortSettings": {
-          "PortName": "/dev/ttyUSB0",
-          "BaudRate": 9600,
-          "DeadPollingTime": 100,
-          "MinimalPollingTimeout": 1000
-        },
-        "ServerAttributes": {
-          "Id": 1,
-          "Name": "UIM-1",
-          "IPAddress": ""
-        }
+        "ServerId": 1,
+        "DeadPollingTime": 100,
+        "PollingTimeout": 10000,
+        "MinimalPollingTimeout": 1000,
+        "ComPortName": "/dev/ttyUSB0",
+        "ComPortBaudRate": 9600
       },
       {
-        "PortSettings": {
-          "PortName": "/dev/ttyUSB1",
-          "BaudRate": 9600,
-          "DeadPollingTime": 100,
-          "MinimalPollingTimeout": 1000
-        },
-        "ServerAttributes": {
-          "Id": 2,
-          "Name": "UIM-2",
-          "IPAddress": ""
-        }
+        "ServerId": 2,
+        "DeadPollingTime": 100,
+        "PollingTimeout": 10000,
+        "MinimalPollingTimeout": 1000,
+        "ComPortName": "/dev/ttyUSB1",
+        "ComPortBaudRate": 9600
       }
     ]
-  },
-  "PortSettings": {
-    "PortName": "/dev/ttyUSB0",
-    "BaudRate": 9600,
-    "DeadPollingTime": 100,
-    "MinimalPollingTimeout": 1000
   },
   "SqlConnectionSettings": {
     "Host": "localhost",
     "Username": "postgres",
     "Password": "postgres",
     "Database": "arms_webapp_development"
-  },
-  "ServerAttributes": {
-    "Id": 1,
-    "Name": "UIM-1",
-    "IPAddress": ""
   }
 }
